Development

+++ hadoop-rasters/src/main/java/vafilonov/hadooprasters/mapreduce/input/AbstractGeoRasterFileReader.java
@@ -1,13 +1,11 @@
 package vafilonov.hadooprasters.mapreduce.input;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Objects;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.json.JsonMapper;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.mapreduce.InputSplit;
@@ -15,9 +13,9 @@ import org.apache.hadoop.mapreduce.RecordReader;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.hadoop.mapreduce.lib.input.FileSplit;
 import vafilonov.hadooprasters.core.util.ConfigUtils;
+import vafilonov.hadooprasters.frontend.model.json.BandConfig;
 import vafilonov.hadooprasters.frontend.model.json.JobInputConfig;
 import vafilonov.hadooprasters.mapreduce.model.GdalDataset;
-import vafilonov.hadooprasters.core.util.JobUtils;
 
 public abstract class AbstractGeoRasterFileReader<KeyType, ValueType> extends RecordReader<KeyType, ValueType> {
 
@@ -25,18 +23,28 @@ public abstract class AbstractGeoRasterFileReader<KeyType, ValueType> extends Re
 
     protected JobInputConfig jobInputConfig;
 
+    protected BandConfig band;
+
+    protected String attemptId;
+
+    protected String localPath;
+
 
     @Override
     public final void initialize(InputSplit split, TaskAttemptContext context) throws IOException, InterruptedException {
-        Configuration conf = context.getConfiguration();
-        jobInputConfig = ConfigUtils.parseConfig(context.getCacheFiles()[0].getPath());
-        Path filepath = ((FileSplit) split).getPath();
-        String fileId = ConfigUtils.getFileIdByPath(filepath.toString(), jobInputConfig);
 
-        String file = ensureLocalPath(filepath, conf, context.getTaskAttemptID().toString());
-        Objects.requireNonNull(file);
-        dataset = GdalDataset.loadDataset(file, context.getJobName());
-        dataset.setFileIdentifier(fileId);
+        attemptId = context.getTaskAttemptID().toString();
+
+        Configuration conf = context.getConfiguration();
+
+        jobInputConfig = ConfigUtils.parseConfig(new Path(context.getCacheFiles()[0]), conf);
+        Path filepath = ((FileSplit) split).getPath();
+        band = ConfigUtils.getBandByPath(filepath.toString(), jobInputConfig);
+
+        localPath = ensureLocalPath(filepath, conf, attemptId);
+        Objects.requireNonNull(localPath);
+        dataset = GdalDataset.loadDataset(localPath, context.getJobName(), band.getBandIndex());
+        dataset.setFileIdentifier(band.getFileId());
 
         innerInitialize((FileSplit) split, context);
     }
@@ -48,7 +56,12 @@ public abstract class AbstractGeoRasterFileReader<KeyType, ValueType> extends Re
 
     @Override
     public void close() throws IOException {
-        dataset.delete();
+        if (dataset != null) {
+            dataset.delete();
+        }
+        if (localPath != null) {
+            Files.deleteIfExists(Paths.get(localPath));
+        }
     }
 
     /**
@@ -64,9 +77,13 @@ public abstract class AbstractGeoRasterFileReader<KeyType, ValueType> extends Re
         if (filePath.toUri().getScheme().equals("file")) {
             localPath = filePath.toUri().getPath();
         } else {
-            Path tempFile = JobUtils.createAttemptTempFile(conf, attemptId);
-            filePath.getFileSystem(conf).copyToLocalFile(filePath, tempFile);
-            localPath = tempFile.toUri().getPath();
+            File f = File.createTempFile(
+                    "temp-geo-",
+                    filePath.getName().substring(filePath.getName().lastIndexOf("."))
+            );
+            //Path tempFile = JobUtils.createAttemptTempFile(conf, attemptId);
+            filePath.getFileSystem(conf).copyToLocalFile(filePath, new Path(f.getAbsolutePath()));
+            localPath = f.getAbsolutePath();
         }
 
         return localPath;
